iot-subscriber/pyproject.toml
[project]
name = "iot-subscriber"
version = "0.1.0"
description = ""
authors = [
    { name = "Seungbae Ji", email = "seungbae.ji@gmail.com" },
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "asyncio-mqtt (>=0.16.2,<0.17.0)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "opentelemetry-api (>=1.34.1,<2.0.0)",
    "opentelemetry-sdk (>=1.34.1,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.34.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aiomqtt (>=2.4.0,<3.0.0)",
]

[tool.poetry]
packages = [
    { include = "iot_subscriber", from = "src" },
]
include = [
    "config/",
]

[tool.poetry.scripts]
iot-subscriber = "iot_subscriber.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
ruff = "^0.4.4"
mypy = "^1.10.0"

[tool.pytest.ini_options]
addopts = "-ra -v"
testpaths = [
    "tests",
]
pythonpath = [
    "src",
]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [
    "tests/data/",
]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
strict = true
mypy_path = "src"

[build-system]
requires = [
    "poetry-core>=2.0.0,<3.0.0",
]
build-backend = "poetry.core.masonry.api"
